test: move setup suite to a single file
